Mesero

+++ src/main/java/com/example/proydbp/client/domain/ClientService.java
@@ -1,27 +1,28 @@
 package com.example.proydbp.client.domain;
 
+import com.example.proydbp.auth.utils.AuthorizationUtils;
 import com.example.proydbp.client.dto.ClientRequestDto;
 import com.example.proydbp.client.dto.ClientResponseDto;
 import com.example.proydbp.client.dto.PatchClientDto;
-
 import com.example.proydbp.client.infrastructure.ClientRepository;
 import com.example.proydbp.delivery.domain.StatusDelivery;
 import com.example.proydbp.delivery.dto.DeliveryResponseDto;
+import com.example.proydbp.exception.UnauthorizeOperationException;
+import com.example.proydbp.exception.UserAlreadyExistException;
 import com.example.proydbp.pedido_local.domain.StatusPedidoLocal;
 import com.example.proydbp.pedido_local.dto.PedidoLocalResponseDto;
 import com.example.proydbp.reservation.domain.StatusReservation;
 import com.example.proydbp.reservation.dto.ReservationResponseDto;
+import com.example.proydbp.reviewDelivery.dto.ReviewDeliveryResponseDto;
+import com.example.proydbp.reviewMesero.dto.ReviewMeseroResponseDto;
 import com.example.proydbp.user.domain.Role;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
-
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.modelmapper.ModelMapper;
-
-
+import java.time.ZonedDateTime;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,13 +30,18 @@ public class ClientService {
 
     final private ClientRepository clientRepository;
     private final ModelMapper modelMapper;
+    private final PasswordEncoder passwordEncoder;
+    private final AuthorizationUtils authorizationUtils;
 
 
     @Autowired
-    public ClientService(ClientRepository clientRepository) {
+    public ClientService(AuthorizationUtils authorizationUtils, ClientRepository clientRepository, PasswordEncoder passwordEncoder, ModelMapper modelMapper) {
         this.clientRepository = clientRepository;
-        this.modelMapper = new ModelMapper();
+        this.modelMapper = modelMapper;
+        this.passwordEncoder = passwordEncoder;
+        this.authorizationUtils = authorizationUtils;
     }
+
 
     public ClientResponseDto getClient(Long id) {
         Client client = clientRepository.findById(id)
@@ -51,32 +57,19 @@ public class ClientService {
     }
 
 
-    public boolean clientExists(ClientRequestDto clientRequestDto) {
-        Optional<Client> existingClient = clientRepository.findByEmail(clientRequestDto.getEmail());
-        return existingClient.isPresent();
-    }
-
-    public boolean clientExists(PatchClientDto patchClientDto) {
-        Optional<Client> existingClient = clientRepository.findByEmail(patchClientDto.getEmail());
-        return existingClient.isPresent();
-    }
-
-
     public ClientResponseDto saveClientDto(ClientRequestDto clientRequestDto) {
+        if (clientRepository.findByEmail(clientRequestDto.getEmail()).isPresent()) {
+            throw new UserAlreadyExistException("Cliente con email " + clientRequestDto.getEmail() + " ya existe.");
+        }
         Client client = modelMapper.map(clientRequestDto, Client.class);
+        client.setPassword(passwordEncoder.encode(clientRequestDto.getPassword()));
         client.setRole(Role.CLIENT);
+        client.setPhoneNumber(clientRequestDto.getPhone());
+        client.setUpdatedAt(ZonedDateTime.now());
+        client.setCreatedAt(ZonedDateTime.now());
+        client.setRango(Rango.BRONZE);
         clientRepository.save(client);
         return modelMapper.map(client, ClientResponseDto.class);
-    }
-
-    public String getIdClient(ClientRequestDto clientRequestDto){
-        Client client = clientRepository.findByEmail(clientRequestDto.getEmail())
-                .orElseThrow(() -> new UsernameNotFoundException("Cliente con email " + clientRequestDto.getEmail() + " no encontrado."));
-        return String.valueOf(client.getId());
-    }
-
-    public boolean clientExists(Long id) {
-        return clientRepository.existsById(id);
     }
 
     public void deleteClient(Long id) {
@@ -85,59 +78,67 @@ public class ClientService {
         clientRepository.deleteById(id);
     }
 
-    public void updateClient(Long id, PatchClientDto patchClientDto) {
+    public ClientResponseDto updateClient(Long id, PatchClientDto patchClientDto) {
         Client client = clientRepository.findById(id)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente con id "+ id +"no encontrado."));
 
-
         client.setFirstName(patchClientDto.getFirstName());
         client.setLastName(patchClientDto.getLastName());
-        client.setEmail(patchClientDto.getEmail());
-        client.setPassword(patchClientDto.getPassword());
-        client.setPhoneNumber(patchClientDto.getPhoneNumber());
+        client.setPassword(passwordEncoder.encode(patchClientDto.getPassword()));
+        client.setPhoneNumber(patchClientDto.getPhone());
+        client.setUpdatedAt(ZonedDateTime.now());
         clientRepository.save(client);
+        return modelMapper.map(client, ClientResponseDto.class);
     }
 
     public ClientResponseDto getAuthenticatedClient(){
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
 
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-
-        Client client = clientRepository
-                .findByEmail(clientName)
-                .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado " + clientName));
+        Client client = clientRepository.findByEmail(username)
+                .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado " + username));
 
         return modelMapper.map(client, ClientResponseDto.class);
-
     }
 
 
     public void deleteAuthenticatedClient(){
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-        Client client = clientRepository
-                .findByEmail(clientName)
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
+
         clientRepository.deleteById(client.getId());
     }
 
-    public void updateAuthenticatedClient(PatchClientDto patchClientDto) {
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-        Client client = clientRepository
-                .findByEmail(clientName)
+    public ClientResponseDto updateAuthenticatedClient(PatchClientDto patchClientDto) {
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
 
         client.setFirstName(patchClientDto.getFirstName());
         client.setLastName(patchClientDto.getLastName());
-        client.setEmail(patchClientDto.getEmail());
-        client.setPassword(patchClientDto.getPassword());
-        client.setPhoneNumber(patchClientDto.getPhoneNumber());
+        client.setPassword(passwordEncoder.encode(patchClientDto.getPassword()));
+        client.setPhoneNumber(patchClientDto.getPhone());
+        client.setUpdatedAt(ZonedDateTime.now());
         clientRepository.save(client);
+        return modelMapper.map(client, ClientResponseDto.class);
     }
 
+
     public List<PedidoLocalResponseDto> getAllPedidoLocal(){
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-        Client client = clientRepository
-                .findByEmail(clientName)
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
+
         return client.getPedidoLocales()
                 .stream()
                 .map(pedidoLocal -> modelMapper.map(pedidoLocal, PedidoLocalResponseDto.class))
@@ -145,10 +146,13 @@ public class ClientService {
     }
 
     public List<DeliveryResponseDto> getAllDelivery(){
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-        Client client = clientRepository
-                .findByEmail(clientName)
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
+
         return client.getDeliveries()
                 .stream()
                 .map(delivery -> modelMapper.map(delivery, DeliveryResponseDto.class))
@@ -156,10 +160,13 @@ public class ClientService {
     }
 
     public List<ReservationResponseDto> getAllReservation(){
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-        Client client = clientRepository
-                .findByEmail(clientName)
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
+
         return client.getReservations()
                 .stream()
                 .map(reservation -> modelMapper.map(reservation, ReservationResponseDto.class))
@@ -167,11 +174,13 @@ public class ClientService {
     }
 
     public Void updateLoyaltyPoints(int point){
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
 
-        Client client = clientRepository
-                .findByEmail(clientName)
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
+
         client.setLoyaltyPoints(point+client.getLoyaltyPoints());
         if (client.getLoyaltyPoints() < 100) {
             client.setRango(Rango.BRONZE);
@@ -185,47 +194,83 @@ public class ClientService {
 
         clientRepository.save(client);
 
-        // Incluir el evento
         return null;
     }
 
 
-    // adicional
 
     public List<PedidoLocalResponseDto> getActualPedidoLocal(){
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-        Client client = clientRepository
-                .findByEmail(clientName)
+
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
+
 
         return client.getPedidoLocales().stream()
-                .filter(pedidoLocal -> pedidoLocal.getStatus() == StatusPedidoLocal.ENTREGADO || pedidoLocal.getStatus() == StatusPedidoLocal.CANCELADO)  // Filtrar por estado enum
+                .filter(pedidoLocal -> pedidoLocal.getStatus() == StatusPedidoLocal.EN_PREPARACION || pedidoLocal.getStatus() == StatusPedidoLocal.LISTO || pedidoLocal.getStatus() == StatusPedidoLocal.RECIBIDO)  // Filtrar por estado enum
                 .map(delivery -> modelMapper.map(delivery, PedidoLocalResponseDto.class))  // Mapear usando ModelMapper
-                .collect(Collectors.toList());  // Recoger la lista filtrada
+                .collect(Collectors.toList());
     }
 
+
     public List<DeliveryResponseDto> getActualDelivery(){
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-        Client client = clientRepository
-                .findByEmail(clientName)
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
-        // Filtrar los deliveries que no están en estado "ENTREGADO" y mapear a DeliveryResponseDto usando ModelMapper
+
 
         return client.getDeliveries().stream()
                 .filter(delivery -> delivery.getStatus() != StatusDelivery.ENTREGADO && delivery.getStatus() != StatusDelivery.CANCELADO)  // Filtrar por estado enum
-                .map(delivery -> modelMapper.map(delivery, DeliveryResponseDto.class))  // Mapear usando ModelMapper
-                .collect(Collectors.toList());  // Recoger la lista filtrada
+                .map(delivery -> modelMapper.map(delivery, DeliveryResponseDto.class))
+                .collect(Collectors.toList());
     }
 
     public List<ReservationResponseDto> getActualReservation(){
-        String clientName = SecurityContextHolder.getContext().getAuthentication().getName();
-        Client client = clientRepository
-                .findByEmail(clientName)
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
 
+
         return client.getReservations().stream()
-                .filter(reservation -> reservation.getStatusReservation() != StatusReservation.CONFIRMADO && reservation.getStatusReservation() != StatusReservation.CANCELADO)  // Filtrar por estado enum
-                .map(delivery -> modelMapper.map(delivery, ReservationResponseDto.class))  // Mapear usando ModelMapper
-                .collect(Collectors.toList());  // Recoger la lista filtrada
+                .filter(reservation -> reservation.getStatusReservation() != StatusReservation.FINALIZADA && reservation.getStatusReservation() != StatusReservation.CANCELADO)  // Filtrar por estado enum
+                .map(delivery -> modelMapper.map(delivery, ReservationResponseDto.class))
+                .collect(Collectors.toList());
+    }
+
+    public List<ReviewMeseroResponseDto> getAllReviewMesero() {
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
+                .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
+
+        return client.getReviewMesero()
+                .stream()
+                .map(reviewMesero -> modelMapper.map(reviewMesero, ReviewMeseroResponseDto.class))
+                .collect(Collectors.toList());
+    }
+
+    public List<ReviewDeliveryResponseDto> getAllReviewDelivery() {
+        String username = authorizationUtils.getCurrentUserEmail();
+        if (username == null)
+            throw new UnauthorizeOperationException("Anonymous User not allowed to access this resource");
+
+        Client client = clientRepository.findByEmail(username)
+                .orElseThrow(() -> new UsernameNotFoundException("Cliente no encontrado"));
+
+        return client.getReviewDelivery()
+                .stream()
+                .map(reviewDelivery -> modelMapper.map(reviewDelivery, ReviewDeliveryResponseDto.class))
+                .collect(Collectors.toList());
     }
 }
